Unify manager/judge assignment/judging areas
