+++ indico/modules/events/papers/templates/_paper_list.html
