+++ indico/modules/events/papers/templates/management/overview.html
@@ -139,7 +139,7 @@
                 {%- trans %}Assign reviewers and judges to papers{% endtrans -%}
             </div>
             <div class="toolbar">
-                <a class="i-button icon-settings" href="{{ url_for('.assignment', event) }}">
+                <a class="i-button icon-settings" href="{{ url_for('.papers_list', event) }}">
                     {%- trans %}Assign papers{% endtrans -%}
                 </a>
             </div>
